Optimize spiHelperArchiveCase (one-click archiving)
Problem:

The current implementation of spiHelperArchiveCase is very inefficient.
It goes section by section, blanking and archiving each one, spending
many seconds on each section. If a particular SPI case has, say, five
closed reports that need to be archived, that means we have to generate
no less than 10 edits to archive the case and spend the better part of
a minute waiting for the script to finish. When SPI is backlogged, these
delays quickly pile up.

Solution:

Instead of going section by section, we can more optimally go through
and do a "one-click archive" by grabbing the case page text only once,
figuring out which lines need to be archived and which lines should be
kept, and then submitting just two edits: one to archive and one to
blank the archived sections. Doing this is tremendously more efficient:
instead of waiting a minute to archive an SPI case that has many
sections, the processing time is now down to just a couple seconds or so.